订单修改价格，增加赠品

+++ src/main/java/com/aiqin/mgs/order/api/domain/SettlementInfo.java
@@ -123,7 +123,7 @@ public class SettlementInfo extends PagesRequest {
      */
     @JsonProperty("shop_order_preferential")
     @ApiModelProperty(value = "店长整单优惠金额")
-    private Long shopOrderPreferential;
+    private Integer shopOrderPreferential;
 }
